fix(JobCard): refactor and cleanup company
