+++ react/JobCard/JobCard.less
@@ -194,11 +194,6 @@
   align-self: flex-start;
 }
 
-.highlight {
-  font-weight: @sa-bold;
-  color: @saBlack;
-}
-
 .positionTitle {
   padding-bottom: @job-card-text-space;
   display: inline;
@@ -221,14 +216,4 @@
   .titleKeyword {
     color: darken(@saVisitedLink, 10%);
   }
-}
-
-.companyName {
-  color: @saGrey1;
-  font-weight: @sa-semi-bold;
-}
-
-.companyLink {
-  &:extend(.companyName);
-  .link(@saGrey1, true);
 }
