+++ react/JobCard/components/LocationGroup/LocationGroup.test.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { shallow } from 'enzyme';
 
 import LocationGroup from './LocationGroup';
-import { Constants } from 'seek-asia-style-guide/react';
 
 const defaultLocations = [
   {
@@ -26,5 +25,5 @@ describe('JobCard - LocationGroup', () => {
   it('should render correctly', () => {
     const wrapper = shallow(<LocationGroup locations={defaultLocations} />);
     expect(wrapper).toMatchSnapshot();
-  })
-})
+  });
+});
